requirements.txt
aioredis==1.3.1
asgiref==3.5.0
async-timeout==4.0.2
attrs==21.4.0
autobahn==22.2.2
Automat==20.2.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.10.0
black==22.1.0
certifi==2021.10.8
cffi==1.15.0
channels==3.0.4
channels-redis==3.4.0
charset-normalizer==2.0.12
click==8.0.4
constantly==15.1.0
cryptography==36.0.2
daphne==3.0.2
Django==4.0.3
django-cors-headers==3.11.0
django-countries==7.3.2
django-filter==21.1
django-phonenumber-field==6.1.0
djangorestframework==3.13.1
flake8==4.0.1
hiredis==2.0.0
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
mccabe==0.6.1
metadata-parser==0.10.5
msgpack==1.0.3
mypy-extensions==0.4.3
pathspec==0.9.0
Pillow==9.0.1
platformdirs==2.5.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pycparser==2.21
pyflakes==2.4.0
pyOpenSSL==22.0.0
python-dotenv==0.20.0
pytz==2022.1
requests==2.27.1
requests-toolbelt==0.9.1
service-identity==21.1.0
six==1.16.0
soupsieve==2.3.1
sqlparse==0.4.2
tomli==2.0.1
Twisted==22.2.0
txaio==22.2.1
typing_extensions==4.1.1
urllib3==1.26.9
zope.interface==5.4.0
